isIntpower {numbers}	R Documentation

Powers of Integers

Description

Determine whether p is the power of an integer.

Usage

  isIntpower(p)

  isSquare(p)
  isSquarefree(p)

Arguments

p
any integer number.

Details

isIntpower(p) determines whether p is the power of an integer and returns a tupel (n, m) such that p=n^m where m is as small as possible. E.g., if p is prime it returns c(p,1).

isSquare(p) determines whether p is the square of an integer; and isSquarefree(p) determines if p contains a square number as a divisor.

Value

A 2-vector of integers.

Examples

isIntpower(1)    #  1  1
isIntpower(15)   # 15  1
isIntpower(17)   # 17  1
isIntpower(64)   #  8  2
isIntpower(36)   #  6  2
isIntpower(100)  # 10  2
## Not run: 
  for (p in 5^7:7^5) {
      pp <- isIntpower(p)
      if (pp[2] != 1) cat(p, ":\t", pp, "\n")
  }
## End(Not run)

[Package numbers version 0.9-2 Index]