bell {numbers}	R Documentation

Bell Numbers

Description

Generate Bell numbers.

Usage

bell(n)

Arguments

n
integer, asking for the n-th Bell number.

Details

Bell numbers, commonly denoted as B_n, are defined as the number of partitions of a set of n elements. They can easily be calculated recursively.

Bell numbers also appear as moments of probability distributions, for example B_n is the n-th momentum of the Poisson distribution with mean 1.

Value

A single integer, as long as n<=22.

Examples

sapply(0:10, bell)
#      1      1      2      5     15     52    203    877   4140  21147 115975

[Package numbers version 0.9-2 Index]