isPrimroot {numbers}	R Documentation

Primitive Root Test

Description

Determine whether g generates the multiplicative group modulo p.

Usage

  isPrimroot(g, p)

Arguments

g
integer greater 2 (and smaller than p).

p
prime number.

Details

Test is done by determining the order of g modulo p.

Value

Returns TRUE or FALSE.

Examples

isPrimroot(2, 7)
isPrimroot(2, 71)
isPrimroot(7, 71)

[Package numbers version 0.9-2 Index]