omega {numbers}	R Documentation

Number of Prime Factors

Description

Number of prime factors resp. sum of all exponents of prime factors in the prime decomposition.

Usage

omega(n)
Omega(n)

Arguments

n
Positive integer.

Details

'omega(n)' returns the number of prime factors of 'n' while 'Omega(n)' returns the sum of their exponents in the prime decomposition. 'omega' and 'Omega' are identical if there are no quadratic factors.

Remark: (-1)^Omega(n) is the Liouville function.

Value

Natural number.

Note

Works well up to 10^9.

See Also

Sigma

Examples

omega(2*3*5*7*11*13*17*19)  #=> 8
Omega(2 * 3^2 * 5^3 * 7^4)  #=> 10

[Package numbers version 0.9-2 Index]