previousPrime {numbers}	R Documentation

Previous Prime

Description

Find the next prime below n.

Usage

  previousPrime(n)

Arguments

n
natural number.

Details

previousPrime finds the next prime number smaller than n, while nextPrime finds the next prime number below n. In general the previousn prime will occur in the interval [n-1,n-log(n)].

In double precision arithmetic integers are represented exactly only up to 2^53 - 1, therefore this is the maximal allowed value.

Value

Integer.

See Also

Primes, isPrime

Examples

p <- previousPrime(1e+6)  # 999983
isPrime(p)                # TRUE

[Package numbers version 0.9-2 Index]