Sigma {numbers}	R Documentation

Divisor Functions

Description

Sum of powers of all divisors of a natural number.

Usage

Sigma(n, k = 1, proper = FALSE)

tau(n)

Arguments

n
Positive integer.

k
Numeric scalar, the exponent to be used.

proper
Logical; if TRUE, n will not be considered as a divisor of itself; default: FALSE.

Details

Total sum of all integer divisors of n to the power of k, including 1 and n.

For k=0 this is the number of divisors, for k=1 it is the sum of all divisors of n.

tau is Ramanujan's tau function, here computed using Sigma(., 5) and Sigma(., 11).

A number is called refactorable, if tau(n) divides n, for example n=12 or n=18.

Value

Natural number, the number or sum of all divisors.

Note

Works well up to 10^9.

References

https://en.wikipedia.org/wiki/Divisor_function

https://en.wikipedia.org/wiki/Ramanujan_tau_function

See Also

primeFactors, divisors

Examples

sapply(1:16, Sigma, k = 0)
sapply(1:16, Sigma, k = 1)
sapply(1:16, Sigma, proper = TRUE)

[Package numbers version 0.9-2 Index]