twinPrimes {numbers}	R Documentation

Twin Primes

Description

Generate a list of twin primes between n1 and n2.

Usage

  twinPrimes(n1, n2)

Arguments

n1, n2
natural numbers with n1 <= n2.

Details

twinPrimes uses Primes and uses diff to find all twin primes in the given interval.

In double precision arithmetic integers are represented exactly only up to 2^53 - 1, therefore this is the maximal allowed value.

Value

Returnes a nx2-matrix, where nis the number of twin primes found, and each twin tuple fills one row.

See Also

Primes

Examples

twinPrimes(1e6+1, 1e6+1001)

[Package numbers version 0.9-2 Index]